+++ internal/checks/remediate.go
@@ -213,7 +213,7 @@ func fixQuarantine(path string) RemediationResult {
 	}
 
 	_ = os.MkdirAll(quarantineDir, 0700)
-	safeName := strings.ReplaceAll(path, "/", "_")
+	safeName := quarantineSafeName(path)
 	ts := time.Now().Format("20060102-150405")
 	qPath := filepath.Join(quarantineDir, fmt.Sprintf("%s_%s", ts, safeName))